+++ stubs/template/database/migrations/2025_01_03_094203_create_pages_table.php
@@ -21,6 +21,7 @@ return new class extends Migration
             $table->datetime('published_at')->nullable();
 
             $table->boolean('menuitem')->default(1);
+            $table->boolean('breadcrumb')->default(1);
 
             $table->json('title');
             $table->json('html_title')->nullable();